Don't jam random globally visible objects into Value type in the compiler.
If we can reference a globally visible object (like a list, or a
threadlocal, or something) from a compiled function, we should treat it
as an arbitrary python object, instead of sticking it into a 'Value'
type, and mapping it as a 'python_free_object'.

Generally, we need to be careful only to try to handle specific
python types where we know it's safe to interact with them in a
'typelike' way. Everything else we should just map as object.