Fix a bug in serialization causing recursive groups to have a problem.
We have to make sure we deserialize all python object bodies
and native types before we actually link everything up in a group,
or else we risk being unable to produce an object for a given
portion of a MutuallyRecursiveTypeGroup.

We were just blindly serializing a staticmethod's function as part
of the method's body, and so if the method wasn't serialized yet
we'd blow up. This change ensures we create the staticmethod
with a None function and set if afterwards, so we can walk the
group and build the bodies before producing the objects themselves.