Release the GIL in a background thread instead of releasing it immediately.
This provides a huge performance improvement when many threads are acquiring
and releasing the GIL simultanously.  Instead of constantly context switching,
we allow the thread that's holding the GIL to lazily release it, and to
reacquire it with very little cost if it turns out that its initial GIL
release was short.