Fix the way we serialize mutually recursive types.
We were writing out the whole type group even when one of the members
is named. This is a problem, because it means that if you serialized native_ast.Type,
you'd get a name, but if you serialized TupleOf(Type), which is unnamed, you'd end
up with the entire definition of everything inside of Type, which is actually quite
a lot of code, and therefore slow to deserialize.

Now, we check if any member of the MutuallyRecursiveTypeGroup is named,
and if so, we serialize that name to stand in for the definition of the entire
type group.

As part of this change, introduce constants

    typed_python.__fully_typed_serialization_version__
    typed_python.__untyped_serialization_version__

which we should increment every time we change the serializaion format, so that
clients can detect when they will have invalid data.