Make sure we can serialize recursive type annotations on functions.
We weren't walking function annotations properly, and we also had
nonsensical handling of forwards (since the annotations often have
forwards in them).

Also, ensure that we serialize forwards in the graph correctly, so that
when we have self-referential functions (who leak Forward objects into their
annotations) we have the same structure on both sides of the wire,
which lets the type identity be stable.

Also, don't insist that the sha-hash of a deserialized group matches whats given.
It's possible that the group depends on something named in the existing codebase,
in which case its sha hash will be different in the source and the destination.