Ensure that the order of items in the MRTG is stable when things have good names.
We want to make sure that if we deserialize a piece of code from a codebase and
all that's changed are things like function code (but names are largely the same)
that everything just works. This means that when we serialize a reference to
a member of a mutually recursive type group, which is going to get identified
by the index of the type in the group, we need to make sure those positions
are stable even when we've made small edits.

This isn't foolproof - at some point, we need to figure out how to identify
if some subset of types from one codebase will be compatible with those
from another.