Ensure that round-trip serialization of an anonymous function is stable.
This is tricky to test because of the memoization we have all over the place,
so we do the serialization the first time in an external process.

We have to be careful to exclude decorators from the syntax trees of
function definitions when we serialize them, since the decorators
(and type annotations) don't actually make their way onto the code objects.