Fix a nasty deserialization bug causing us to create segfaulty functions.
Basically, functions at module scope that were recursive were encoding their
reference to themselves as a 'cell', when it should be through the globals. The
resulting object would look up in the cells array and not find it and blow up.
This only comes up if you have a module-level object thats not serializable
by name, which doesn't happen most of the time, but can happen if you're
evaluating code using 'eval'.