Ensure that OneOf instances and method calls work.
To really get this right throughout, we'll have to tighten up our
model for how masquerades and oneofs interact. It should be possible
to trace the masquerade through a one-of, but right now we just
jam it down to the interpreter type representation.

Still, for most cases, everything should just work, and even
if we lose typing, we should still be able to compile.