Make sure to tag along names of full modules serialized in function globals.
This basically adds an API that allows us to insist that modules we need
for a function to execute are ensured imported when we deserialize a function.

If a function's globals dictionary is not a proper module and we serialize it,
we always check which of the globals are actually used. This change ensures
that if you use 'numpy' and the globals contains 'numpy.whatever', that
'numpy.whatever' gets tracked along with it.

Clients can use this to encode the submodules of a particular module they
need for the function to execute.

This is used in other projects where we have client code that looks like

    import module.submodule

    def f():
        module.submodule.something()

and then we serialize 'f'.  Our other project ensures that 'module.submodule'
is encoded in f's globals, so that when f gets deserialized on another machine
we can assume that submodule is imported.