Make sure we can compile against OneOf Alternative.matches.
In general, the type hierarchy needs to be fully modeled in C++.

This means that for each distinct type we want to model in the compiler,
we need an explicit type in the interpreter layer to keep track of it.
We can then have 'interpreter types masquerading as typed objects' as
a second layer. But there's no getting around the idea that the
typeRepresentation for a given Wrapper object needs to uniquely
identify the Wrapper.