Fix a bug causing invalid variable accesses in compiled code.
When the compiler knows a variable 'X' is OneOf(A, B), but then determines
that it can, in some pathway, only be a 'B', it accesses 'X' directly without
checking the flag (for performance).  Unfortunately, if instances of 'A' can
also be instances of 'B', then its possible for values to get encoded as
'A', even though the compiler knows it as 'B'.

We protect against this by checking whether if we know that a OneOf(A, B)
is definitely a 'B' that that means it is definitely not an A. In this
case we can omit the check, but otherwise we need to explicitly
perform the cast.