Improve compiler performance.
Mostly, by moving some very verbose operations into their own function
definitions. For instance, we were explicitly generating the code
for OneOf copy construction/deletion all over the place, which is
very expensive compared to placing it in a function.

Also, as part of this change, don't use the function-pointer based
dispatch mechanism on class destructors if the class is Final.