Rework "held classes".
'Held' classes can now be instantiated by the interpreter, and
returning references to them from their storage context returns 'a temporary'
reference to them. RefTo is now deprecated.

This gives us the ability to write very fast classes that don't use refcounting,
at the expense of somewhat more complex lifetime semantics. Speficially, if H is
a held class and you write

    x = H()
    y = x

then 'x' will be a copy of 'H'. In the interpreter, 'y' will reference 'x',
but in the compiler it may be a copy. The compiler is free to introduce duplicates of
any HeldClass that's not explicitly allocated inside of a Class, ListOf, etc.

If l is a ListOf(H), then l[i] is a 'temporary' reference to the HeldClass instance
at l[ix] and will exist as a reference for the duration of the expression. This
allows expressions like

    l[ix].x += 1

to work (because l[ix] is a temporary reference) but ensures that

    x = l[ix]

is a copy of l[ix], so that you don't run the risk of accessing a deallocated
element (say, because l gets deallocated).