Don't allow serialization to depend on whether an object is in a mutually recursive type group.
Before, if the compiler added a function to an MRTG, we'd see that in serialization and then
serialize the type group and then a memo. This is the correct behavior, except that it makes
serialization dependent on whether something has had its type identity computed, which is
bad because then serialization is not stable.

The solution is to only allow us to _start_ serializing a mutually recursive type group
when we're serializing a Type or a python class, but not a regular function.