Implement function signatures. Fixes #313. Fixes #312. Fixes #291.
This change implements 'function signatures', which allow us to
express the idea that a function's stated return type can depend
on the types of its arguments in a way that is more complex
than simple overloading.

As part of this change, we also rework how subclasses can change
the return type annotations when overriding base-class members,
and introduce the constraint that classmembers cannot be overridden
by base classes (more on why this is a good idea below).

The core idea implemented in this change is to allow f
unctions to use a 'signature function' instead of a regular return
type annotation. A signature function is any non-type callable object.
An example is:

    @Function
    def f(x) -> lambda X: someTypeExpressionOfX:
        ... function definition ...

In this case, instead of the function's return type being a
single type specified at parse time, it's a function of the
type of x.  More generally, the types of the positional and
keyword arguments to 'f' will be passed to the signature
function when the function is invoked, and the return type
will be taken as the return type of the function.

The primary use-case for this is to be able to declare more
sophisticated type relationships inside of classes. As part
of this change, we also formalize the relationship between
the function signatures of base and child classes. Prior to
this change, all methods defined in base classes that were not
marked 'Final' had to have a return type annotated, and all
child classes had to exactly match that signature.

Now, the rule is that the concrete return type of any method
in a child class must be 'covered by' (that is, trivially convertible
to - int -> OneOf(int, float), for instance) the return type given
by matching that same set of arguments in the base class.

Because type signature functions can be arbitrarily complicated,
this constraint is enforced at run-time, since we can't forsee
exactly how all combinations of arguments might play out at
compile / function definition time.

This is a necessary condition to allow the compiler to generate
efficient code when the compiler knows the type of the base
class: if base class B has method 'f', and child class C overrides
it, we need for the compiler to assume that if B.f returns T,
then so will _all subclasses_. Otherwise, the compiler can't
make any reasonable type inferences whatsoever.

By way of example, this means that if you write

    class B(Class):
        def f(self, x: int) -> OneOf(int, str):
            ....

then a child class may override 'f' to return 'int', for example,
but not 'float', since the compiler will assume that B.f (when
matched with ints) will always return an int or a string. Subclasses
are free to override it with other types for other argument
signatures that wouldn't match the ones defined in the
base class.

There are some difficult subtleties associated with signature functions
and classes: if you have a class 'B' that's not marked Final, then
a type annotation (x: B) indicates that 'x' is B or a subclass of
B. In this case, the compiler cannot pass 'B' itself to the signature
function, because it doesn't actually know the concrete type of B.

As an example, imagine you wrote

    @Function
    def f(x: B) -> lambda X: ListOf(X):
        return [x]

in this case, if you were to pass an instance of 'C' (a child class
of B), then you'd get a materially different concrete type as
a result.

This completely breaks the compiler's ability to reason about the
code you're writing.

The solution we've adopted is to, in this case, pass an object
called 'SubclassOf' (e.g. SubclassOf(B)) which gives you access
to any classmembers defined in B. In order for this to be sensible,
we've also implemented a constraint where if 'B' defines classmember
'X', then child classes cannot override it. This allows us to write
meaningful signature functions that are 'covariant' (meaning that
as we make the argument types more specific, the signature we
deduce also gets more specific).