Add 'SubclassOf' for holding Class type objects.
This is a step towards a more formal model of classmethod inheritance
that can be used 'virtually'. Currently, if you define a classmethod
'f' in a class that not final, you can't use it in the compiler because
the compiler can't know if you've overriden it.

The existing compiler implementation just kind of punted on that.

The new implementation now knows enough to track Type* objects
through native code, and so the next natural step is to allow
the same rules we have for instances to apply to staticmethods
and classmethods, so that the compiler can generate efficient
code.