Implement virtual dispatch for Class type instances.
As a result of this change, if you declare a static or
classmethod on a Class, and override it on the base, then
the same rules of type restriction and inheritance apply
as with regular instances. This means that if you have an
instance of a type object and you access a classmethod,
the compiler can use the same tricks it uses with instances
to make fast dispatch code.

As a result, you can now access class and staticmethods on
type objects and instances even if the compiler can't prove
the exact instance type of the class.

We still have yet to implement this correctly for __init__,
which also should basically have this same behavior.

Also, it's currently undefined if you ask the same method
to be a classmethod in a base class and a staticmethod
in a childclass.