Fix our pyproject.toml to deal with the numpy build issue.
Basically what happens is this: pyproject.toml specifies what versions of
each library we need to build against. The requirements.txt file describes
what versions of each library we need at runtime.

When we install typed_python, pip picks a version of numpy to build against.
In most cases, this will be the same version of numpy that's going to be
installed.  However, pip doesn't _HAVE_ to do this - it's allowed to
pick any python version that matches your requirements file. In particular,
if some other dependency has pinned a specific version of numpy, you may
end up building against the newest and being installed against an older one.

This turns out to be a problem because when you compile against a specific
version of numpy, you depend on the binary layout of the objects, and newer
versions of numpy are not backward compatible (at the binary level).

There is a long thread about this issue here:

    https://github.com/pypa/pip/issues/9542

Essentially, the problem is that there is no way to tell pip that the
version it builds against must be identical to the version that it
finally deploys against.  The thread has a great deal of hand-wringing
around why this is so impossible to get right.

It turns out that at least in this case, there is a workaround for this
problem: numpy versions are forward compatible - meaning that if you
build against an older version of numpy, you will be able to run against
newer versions of numpy.  As a result, in pyproject.toml, you can specify
'oldest-supported-numpy' as a build requirement, which will force it
to build against an old numpy, which should work with any numpy it
can install.