Serializer uses a _types function to build code objects instead of the builtins.
In particular, calling types.CodeType wasn't maintaining the identity of
the tuples that get attached to the resulting code object. This means that
a deepcopied function is subtly different than its original counterpart
and in particular that the sha-hash of a Code object wasn't invariant
over deepcopy which causes shenanigans higher up in the stack.