Ensure that held class temporaries get resolved over multiline expressions correctly.
Basically, we want to ensure that we can write expressions like

	listOfHeldClass[0].x = 10

and have the assignment modify the copy of the held class that lives within the list.
To do this, we need for listOfHeldClass[0] to return a RefTo(H) (where H is whatever
the list is holding). Unfortunately, that would mean that we have a reference to a held
class living in the python heap, which is very likely to cause a segfault, and which
doesn't mirror the semantics of the compiler.

The solution we adopted originally was to keep track of all held class temporaries
and to convert them to non-temporaries (e.g. make copies of the object) when the
current statement has finished executing.  This mirrors the way the compiler works:
temporaries exist for the duration of the expression we're compiling, which allows
chained sequences like this to work.

HOWEVER, the mechanism we use is to put a trace on the parent stackframe, and to
use that trace to discover when we've hit the next frame. Unfortunately, an expression like

	listOfHeldClass[0].f(
		10
	)

causes us to evaluate the '10' on a new line, which triggers the temporary to resolve. This
breaks the entire model.

This change causes us to cache the subset of line numbers that are the starts of new statements.
These line numbers are the ones where the references should be converted to concrete
instances of the held class.