Ensure that we don't arbitrarily duplicate HeldClass instances when calling functions.
In general, we want to ensure that if we pass a held class into a function, and
that held class has a clear storage owner in the parent context, that we pass
a reference to it into the function, so that we can modify the instance.

For that to work, we need to not be making a bunch of copies of it.