Ensure that Held Class types can reference themselves.
This fixes a nasty serialization bug where mutually recursive
type groups can hold references to the same HeldClass in two places
(once as a pyobj and once as a type). This was causing serialization
of held classes that refer to themselves in their own functions
to break.

Really, we should ensure that TypeOrPyObj has a single way
of referring to any given object - there shouldn't be a distinction
in a MutuallyRecursiveTypeGroup between the Type and the
PyObject representation of it.