StringType::cmpStatic: fix accidental reference to std::right
Because of the `using namespace std` directive at the top of
StringType.cpp, within `StringType::cmpStatic`, `right` is
resolved to the function `std::right`, from the <ios> header,
included indirectly by <iostream>. `std::right`, a function,
is always non-null. The intention here was clearly to use
the parameter `right_data`, not `std::right`.

Fix that bug.