Make sure we throw an exception if we attempt to hash an MRTG more than once.
Somehow, we're seeing cases where the graph appears to be unstable and so
we get cycles in the MRTGs which is not supposed to happen. In the future,
we need to implement something else to ensure that if this happens
we can collapse the groups. But really they're supposed to be stable,
so we should try to figure out why the python object graph is not stable.