More fixes for type identities.
It turns out that we were walking into arbitrary python objects. Our compiler doesn't,
in fact, do that.  For instance if you have

    class C:
        def __init__(self, x):
            self.x = x

    c = C(10)

    @Entrypoint
    def f():
        return c.x

the compiler knows that 'x' might change and will not inline the constant. Unfortunately, our
identityHasher was walking in and including the '10' in the value of the hash, which can lead to
serious instability.