Ensure that regular classes get into the type-function memo.
There's a tricky bug where SerializationContext memoizes types based on their
type identity.  Its very important that any class that gets serialized that's
produced by a TypeFunction gets serialized by noting the type function and
its arguments rather than by serializing the class itself. Otherwise, depending
on whether you have called the TypeFunction or deserialized an instance of it,
the memo inside the serialization library can have different versions of the
type, which leads to all kinds of chaos.

This bug exposes another possible bug, where if a type function defines
two types internally, only one of them will be registered by the type
function memo. This is something we need to think about more deeply to really
fix. For now the rules for type functions are:

TypeFunction can either return a type that is primitive, a type that is
a compound of primitives (e.g. OneOf, NamedTuple, Tuple, ListOf, etc),
a type that is named by looking at __module__ and __name__ and finding
that type in that location, or by returning a single class type, inheriting
from object or Class respectively.