Rewrite the MRTG framework, and fix some associated serialization issues.
Prior to this change, the MRTG (MutuallyRecursiveTypeGroup) framework was not threadsafe.
What was happening was that if two threads both attempted to hash a big group of types,
such as a large class with many functions in it, there was a race condition where one
thread would check in the entire group while the other was still traversing it. The
other group would then construct a second group consisting of the types it had visited,
but none of the ones that it had not seen but that had just been installed, and this
could lead to two copies of the type existing with different hashes (since this can
happen during deserialization of type groups).

This change fixes that by rewriting the entire hashing and MRTG framework to have
a disciplined model for how threads interact while they are hashing the same
objects and types, and for how our framework walks the python and Type* graph.

Prior to this change, we had two different ways of walking the graph for hashing
purposes - one for discovery (CompilerVisibleObjectVisitor) and one for hashing
(Type::computeIdentityHash).  Now, we have one way of walking objects, which is
owned by CompilerVisibleObjectVisitor, who invokes a Type's visitCompilerVisibleInternals
method to walk the internals.  This allows for a single unified framework where
we are guaranteed that the same object relations that are used to determine the
graph structure of the MRTG are also used in hashing. We also ensured that we
don't have any Instance objects floating around in this framework - everything
is either a PyObject* or a Type*, encapsulated by instances of TypeOrPyobj.

Finally, we correct a number of serious errors in the serialization/deserialization
layer that were sliding undetected under the prior framework. In particular, we're
careful to preserve the "Forward" structure (since Forward instances can sometimes
make their way into modules), and to be sure that we are consistent about how
we call 'representationFor' and 'nameForObject' when serializing, for both
Type* and PyObject*.  We also introduce SerializationContext.factoryFor to
handle TypeFunction classes, since these, like nameForObject, can be uniquely
identified by a much simpler-to-deserialize set of objects.