Fix a couple of bugs causing serialization/hashing to not be stable.
In particular, make sure we don't depend on __file__ for anything,
since that would make it impossible to relocate a module on disk
without affecting it. We also make sure that the function annotations
dict is kept in sync with the relevant FunctionType* when we
complete a Forward type cycle - otherwise depending on whether we
look at python or C++ we get a different answer.