Use the identityHash of a TypeFunction's arguments in its lookup.
We can't rely on the (python) identity of two Function types being the same
even if they are the same - this happens when we deserialize a type before
we actually create it at the module level (which can happen in the compiler
cache) in which case we can have two python objects representing the same
type. At the moment we don't have a good way of deduplicating these,
so instead, we need to make sure that we use type-identity-hash instead
of python identity/hash in dictionaries involving types.