Ensure that ModuleRepresentation preserves the object graph when copying objects from two different upstream modules.
A ModuleRepresentation models a python module and a collection of objects that can "see"
each other through that module.  It lets you duplicate a collection of such objects
from one module to another so that you can mutate the new copy's module dict and preserve
the old one.

Its important that we maintain the relationships of the objects in the module graph
that may have been copied through different modules. For instance, if you have

    class A:
        ...

and later define

    class B(A):
        ....

then we want to maintain the invariant that in any ModuleRepresentation we get,
"A"'s identity is preserved. We will always have different copies of A (one per
ModuleRepresentation) so we can't use object identity. So instead, we keep track
of the "source module" where an object was first created and use that as a proxy
for its identity.  That way 'A' and 'B' can get copied into different module representations
and if they then get copied back into a single common case, we're guaranteed that
we have only one copy of 'A' and that B's base-class will be the 'A' that's in
the module (an invariant we were not maintaining before this change).