Rewrite much of ModuleRepresentation to fix a bug.
Basically, we weren't being careful about the difference between members
of a module that we are going to modify and therefore need to be deepcopied
(here called 'active') and those that we simply wanted to define but
not deepcopy (here called 'inactive').

This led to problems where DAGs of ModuleRepresentations could not
tell the difference between objects that are 'external' (not involving
the module at all) and those that simply were imported from a different
copy of the module. As a result, we'd end up with all kinds of weird
copies of things.

This change fixes that. For more, see ModuleRepresentation.hpp.