Make the new ModuleRepresentation more precise.
In particular, make sure that all visible values in upstream
modules are considered 'inactive' if they are not otherwise specified.

This solves the problem where a base-class may be duplicated by accident.
Consider the following case:

m1:
    class B:
        pass

m2:
    class C(B):
        def f(self):
            return g

m3:
    def g(): ...

if, when we copy objects from m2 to m3, we fail to include 'B'
as an inactive value, we'll get a separate copy of it. So instead
we make sure that _all_ upstream module values get copied in as
inactive if they are not otherwise mentioned.