Ensure we hold the runtime_lock when deserializing.
We had some deadlocks in other systems coming from holding this lock,
but it turns out that importing modules outside of this lock creates
serious problems. Notably, two threads can attempt to deserialize
something that requires importing a module - the first one starts
importing the module and the second one sees the module entry in
sys.modules and then uses a partially imported module object,
which obviously will create huge problems since the module
member named in the serialized byte stream may not be present.

In order to ensure consistent modules, we have to have importing
happen under the runtime lock. For this to not deadlock, its
very important that all clients who call into SerializationContext
from a module import do so outside of any other lock! Otherwise
we can get a deadlock.