Refactor timestamp module.
This commit adds an explicit model of DateTimes/Timezones, in
many ways mirroring the standard datetime python module.

To expound the theory motivating these changes, we will approximate
the Earth as an inertial frame. In this frame, we can ascribe to
every event a unique set of geographical coordinates *and* a number
representing its time, which we take to be the UTC timestamp.

The locus of events with common timestamp, so defined,
collectively denote an “instant” in time.

There is a bijection between instants in time and UTC timestamps.

By contrast, no such bijection need exist between local clock readings
and instants in time. For example, when we "spring forward" in
daylight savings time, some clock readings-- like 2:30a on
13 March 2022 in NYC-- do not exist; and when we "fall back", some
clock readings-- like 1:30a on 6 November 2022-- exist twice.

The additional types defined in this commit provide a clear API
for identifying unique instants of time.

For some examples:
- the program will raise if you try to create
timestamps from clock readings that do not exist, e.g.,
`NYC.timestamp(DateTime(2022, 3, 13, 2, 30, 0))` raises
a NonExistentDateTime exception.

- to identify the second occurrence of 1:30a 6 Nov 2022 in NYC, you
can call `NYC.timestamp(DateTime(2022, 11, 6, 1, 30, 0), afterFold=True)`.