Ensure we can compile Dict(object, T).
We actually have a bit of a problem with the TP hashing infrastructure
because it's not invariant to the known type. For instance, a Dict(object, T)
will hash a 'float' using python's builtin 'hash' function, but a Dict(float, T)
will use the TP hash function. This is problematic because we now have to be
very careful when we use the word 'hash' whether we're using the python
built-in hash function, or the typed-python hash function indexed by some type
T.

It would be better to just unify the representation of 'hash' throughout
and make it tie out exactly with python.