Force compilation of generator __fastnext__ on instantiation.
One problem with our 'class' virtual inheritance model is that
we don't know what methods we'll need when the class is first
instantiated. For instance, if you write

    class C(Class):
        def f(self, x):
            return str(x)

then we actually have to be prepared to instantiate 'C' with arbitrary
argument types for 'x'.  Practially, this plays out by having 'C' have
an allocated 'slot' for each distinct type that 'x' takes in the compiler,
and we generate the instantiation for a specific child class B's use of
a given type for 'x' when we first use it.

If we didn't do this, then we'd have to compile the full cross product
of all subclasses of C and all possible types ever used for 'x', which
could be very large.

Unfortunately, this can create strange-looking performance problems
(a version of the hot-cache problem) since even if we compile code
that uses both 'C' and 'B', we may not have compiled the dispatch
until it actually gets used.

This change builds a simple framework for allowing generators to declare
that they want their __fastnext__ and destructor compiled as soon as
they are created (in a compiled context). This is reasonable because
those methods take no arguments.

Eventually, we might want to consider immediately compiling all such
cases (where there is no type variation) since this is what we would
do if we were in a more traditional setting (say, C++).