Fix a compiler bug in implicit conversion of ListOf to TupleOf.
We have this confusing thing where function signature types can change
as the compiler gets more information. When we first encounter a function
call, we might return 'None' from the 'convert_call' operation , which means that
control flow does not return.  This is not the same as getting the typeWrapper
for NoneType, which means it actually might return control flow with the
value of 'None'. Instead it means that that particular invocation of
the function doesn't return at all, and we don't have a defined return
type.  This is a problem in convert_to_(type|self)_with_target calls
because those calls are expected to return a bool with success or failure.

This change makes us robust to the intermediate state where the compiler
tells us that we'll throw exceptions.