Strongly-typed 'kwargs' masquerade admits an 'items' method.
Iterating an 'items' method still ends up squashing the keyword arg types
into a giant OneOf, despite the fact that we know the types during the
iteration. We could improve this by figuring out which variables are
active in the target of a 'get_iteration_expressions' and explicitly
activating the specific subtypes for each unrolled level of the loop.