Prune the graph of inflight functions to not include the ones we don't need.
When we first call a python function 'f' with a specific set of arguments, we may not know
its return type the first time we try to convert it.  To ensure we have a stable typing graph,
we repeatedly update the active functions in our graph until the type graph is stable.

This can lead to many copies of the same function, or even multiple signatures of the
same function, only one of which we'll use.

This change prunes those away before we submit them to the LLVM layer.