Add __convert_from__ to Class objects to allow implicit conversion, and switch on for generators. Fixes #416.
Currently, this only works in the interpreter - there's no compiler layer for it.

It's intended to solve the problem where if you write

    class C(Class):
        def __iter__(self, x) -> Generator(int):
            yield 1

then in interpreted code you'll end up returning a regular iterator, not a
Generator, and it just breaks. This change allows the 'generator' to get automatically
converted to a Generator(T).

We're only using this here - we may not keep this feature, since it introduces a whole
host of questions about what it means to 'convert' from a type to a class. But for now
it solves the problem.